+++ vue/src/Components/Chat/ChatMessage.vue
@@ -2,16 +2,20 @@
   <li :class="chatResponseClasses">
     <div class="ai-chat-response-content-wrapper">
       <div class="ai-chat-response-body">
-        <Markdown :markdown="message.content"/>
-        <span v-if="isStreaming" class="streaming-cursor">▊</span>
+        <Markdown :markdown="message.content" />
+        <span
+          v-if="isStreaming"
+          class="streaming-cursor"
+        >▊</span>
       </div>
     </div>
   </li>
 </template>
 
 <script lang="ts">
-import { defineComponent } from 'vue';
+import { defineComponent, PropType } from 'vue';
 import Markdown from '../Markdown.vue';
+import { Message } from '../../types';
 
 export default defineComponent({
   components: {
@@ -19,7 +23,7 @@ export default defineComponent({
   },
   props: {
     message: {
-      type: Object,
+      type: Object as PropType<Message>,
       required: true,
     },
     aiName: {
@@ -63,10 +67,6 @@ export default defineComponent({
       padding: 6px 16px;
       border-radius: 16px;
     }
-  }
-
-  &.ai-chat-assistant-response {
-
   }
 
   .ai-chat-response-content-wrapper {